% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chart_age}
\alias{chart_age}
\title{Age by sex (pyramid plot)}
\usage{
chart_age(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by \link{create_op_all}}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory}
}
\value{
Age by sex pyramid plot in PNG format saved in the current working
    directory or in a specified directory if \code{filename} is a path.
}
\description{
A wrapper function to the \link{pyramid.plot} function to create an age by
sex pyramid plot
}
\examples{
  # Create age by sex pyramid plot using indicators.ALL dataset
  chart_age(x = indicators.ALL,
            filename = paste(tempdir(), "TEST", sep = "/"))

}
