% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chart_csid}
\alias{chart_csid}
\title{Chart dementia screen (CSID) indicators}
\usage{
chart_csid(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by \link{create_op_all}}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory}
}
\value{
Bar plot of CSID in PNG format saved in current working directory
  or in a specified directory if \code{filename} is a path.
}
\description{
Chart dementia screen (CSID) indicators
}
\examples{
  # Create chart using \code{indicators.ALL} dataset
  chart_csid(x = indicators.ALL,
             filename = paste(tempdir(), "TEST", sep = "/"))

}
