% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chart_hhs}
\alias{chart_hhs}
\title{Chart household hunger scale (HHS) indicators}
\usage{
chart_hhs(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by \link{create_op_all}}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory}
}
\value{
Bar plot of HHS in PNG format saved in current working directory
  or in a specified directory if \code{filename} is a path.
}
\description{
Chart household hunger scale (HHS) indicators
}
\examples{
  # Create chart using \code{indicators.ALL} dataset
  chart_hhs(x = indicators.ALL,
            filename = paste(tempdir(), "TEST", sep = "/"))

}
