% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chart_income}
\alias{chart_income}
\title{Chart income indicators}
\usage{
chart_income(x.male, x.female, filename)
}
\arguments{
\item{x.male}{Male subset of indicator dataset}

\item{x.female}{Female subset of indicator dataset}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Bar chart of sources of income by sex in PNG format saved in current
  working directory or in a specified directory if \code{filename} is a path.
}
\description{
Chart income indicators
}
\examples{
  # Create chart using \code{indicators.FEMALES} and \code{indicators.MALES}
  # dataset
  chart_income(x.male = indicators.MALES,
               x.female = indicators.FEMALES,
               filename = paste(tempdir(), "TEST", sep = "/"))

}
