% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chart_k6}
\alias{chart_k6}
\title{Distribution of K6 (overall and by sex)}
\usage{
chart_k6(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by \link{create_op_all}}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory}
}
\value{
Histogram of K6 score in PNG format saved in current
    working directory or in a specified directory if \code{filename} is a
    path.
}
\description{
Distribution of K6 (overall and by sex)
}
\examples{
  # Create chart using \code{indicators.ALL} dataset
  chart_k6(x = indicators.ALL,
           filename = paste(tempdir(), "TEST", sep = "/"))

}
