% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chart_muac}
\alias{chart_muac}
\title{Distribution of MUAC (overall and by sex)}
\usage{
chart_muac(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by \link{create_op_all}}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory}
}
\value{
Histogram of MUAC distribution in PNG format and saved in the current
    working directory or in a specified directory if \code{filename} is a
    path.
}
\description{
Distribution of MUAC (overall and by sex)
}
\examples{
  # Create MUAC histogram using indicators.ALL dataset
  chart_muac(x = indicators.ALL,
             filename = paste(tempdir(), "TEST", sep = "/"))
}
