% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chart_wg}
\alias{chart_wg}
\title{Chart disability (Washington Group - WG) indicators}
\usage{
chart_wg(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by \link{create_op_all}}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory}
}
\value{
Bar plot of Disability Score in PNG format saved in current working
    directory or in a specified directory if \code{filename} is a path.
}
\description{
Chart disability (Washington Group - WG) indicators
}
\examples{
  # Create chart using \code{indicators.ALL} dataset
  chart_wg(x = indicators.ALL, filename = paste(tempdir(), "TEST", sep = "/"))

}
