% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chartADL}
\alias{chartADL}
\title{Distribution of ADL (overall and by sex)}
\usage{
chartADL(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Bar plot of ADL in PNG format saved in current working directory
}
\description{
Distribution of ADL (overall and by sex)
}
\examples{
  # Create chart using \code{indicators.ALL} dataset
  chartADL(x = indicators.ALL, filename = paste(tempdir(), "TEST", sep = "/"))
}
