% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chartAge}
\alias{chartAge}
\title{Age by sex (pyramid plot)}
\usage{
chartAge(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Age by sex pyramid plot in PNG format saved in the current working
    directory
}
\description{
Age by sex (pyramid plot)
}
\examples{
  # Create age by sex pyramid plot using \code{indicators.ALL} dataset
  chartAge(x = indicators.ALL, filename = paste(tempdir(), "TEST", sep = "/"))

}
