% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chartCSID}
\alias{chartCSID}
\title{chartCSID}
\usage{
chartCSID(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Bar plot of CSID in PNG format saved in current working directory
}
\description{
Dementia screen (CSID)
}
\examples{
  # Create chart using \code{indicators.ALL} dataset
  chartCSID(x = indicators.ALL, filename = "TEST")
}
