% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-chart.R
\name{chartDDS}
\alias{chartDDS}
\title{chartDDS}
\usage{
chartDDS(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Barplot of dietary diversity score in PNG format saved in current
    working directory
}
\description{
Distribution of DDS (overall and by sex)
}
\examples{
  # Create DDS chart using \\code{indicators.ALL} dataset
  chartDDS(x = indicators.ALL, filename = "TEST")
}
