% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chartHHS}
\alias{chartHHS}
\title{chartHHS}
\usage{
chartHHS(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Bar plot of HHS in PNG format saved in current working directory
}
\description{
Household hunger scale (HHS)
}
\examples{
  # Create chart using \\code{indicators.ALL} dataset
  chartHHS(x = indicators.ALL, filename = "TEST")
}
