% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-chart.R
\name{chartIncome}
\alias{chartIncome}
\title{chartIncome}
\usage{
chartIncome(x.male, x.female, filename)
}
\arguments{
\item{x.male}{Male subset of indicator dataset}

\item{x.female}{Female subset of indicator dataset}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Bar chart of sources of income by sex
}
\description{
chartIncome
}
\examples{
  # Create chart using \\code{indicators.FEMALES} and \\code{indicators.MALES}
  # dataset
  chartIncome(x.male = indicators.MALES,
              x.female = indicators.FEMALES,
              filename = "TEST")
}
