% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chartK6}
\alias{chartK6}
\title{Distribution of K6 (overall and by sex)}
\usage{
chartK6(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Histogram of K6 score in PNG format saved in current
    working directory
}
\description{
Distribution of K6 (overall and by sex)
}
\examples{
  # Create chart using \code{indicators.ALL} dataset
  chartK6(x = indicators.ALL, filename = paste(tempdir(), "TEST", sep = "/"))
}
