% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chartMF}
\alias{chartMF}
\title{chartMF}
\usage{
chartMF(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Barplot of meal frequency in PNG format saved in current working
    directory
}
\description{
Distribution of meal frequency (overall and by sex)
}
\examples{
  # Create meal frequency chart using \code{indicators.ALL} dataset
  chartMF(x = indicators.ALL, filename = "TEST")
}
