% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chartMUAC}
\alias{chartMUAC}
\title{chartMUAC}
\usage{
chartMUAC(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Histogram of MUAC distribution in PNG format and saved in the current
    working directory
}
\description{
Distribution of MUAC (overall and by sex)
}
\examples{
  # Create MUAC histogram using \code{indicators.ALL} dataset
  chartMUAC(x = indicators.ALL, filename = "TEST")
}
