% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chartWG}
\alias{chartWG}
\title{Chart disability (Washington Group - WG) indicators}
\usage{
chartWG(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Bar plot of Disability Score in PNG format saved in current working
    directory
}
\description{
Chart disability (Washington Group - WG) indicators
}
\examples{
  # Create chart using \code{indicators.ALL} dataset
  chartWG(x = indicators.ALL, filename = paste(tempdir(), "TEST", sep = "/"))
}
