% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{check_indicators}
\alias{check_indicators}
\title{Check whether indicators are RAM-OP indicators}
\usage{
check_indicators(indicators)
}
\description{
Check whether indicators are RAM-OP indicators
}
\keyword{internal}
