% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_adl}
\alias{create_op_adl}
\title{create_op_adl}
\usage{
create_op_adl(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of older people indicators on adl
}
\description{
Create older people indicators dataframe on adl from survey data
collected using the standard RAM-OP questionnaire
}
\section{Indicators}{
 Katz "Index of Independence in Activities of Daily Living" (ADL) score

The Katz ADL score is described in:

\cite{Katz S, Ford AB, Moskowitz RW, Jackson BA, Jaffe MW (1963). Studies
of illness in the aged. The Index of ADL: a standardized measure of
biological and psychosocial function. JAMA, 1963, 185(12):914-9
\url{doi:10.1001/jama.1963.03060120024016}}

\cite{Katz S, Down TD, Cash HR, Grotz, RC (1970). Progress in the development
of the index of ADL. The Gerontologist, 10(1), 20-30
\url{doi:10.1093/geront/10.4_Part_1.274}}

\cite{Katz S (1983). Assessing self-maintenance: Activities of daily living,
mobility and instrumental activities of daily living. JAGS, 31(12),
721-726 \url{doi:10.1111/j.1532-5415.1983.tb03391.x}}

\describe{
\item{\code{ADL01}}{Bathing}
\item{\code{ADL02}}{Dressing}
\item{\code{ADL03}}{Toileting}
\item{\code{ADL04}}{Transferring (mobility)}
\item{\code{ADL05}}{Continence}
\item{\code{ADL06}}{Feeding}
\item{\code{scoreADL}}{ADL Score}
\item{\code{classADL1}}{Severity of dependence 1}
\item{\code{classADL2}}{Severity of dependence 2}
\item{\code{classADL3}}{Severity of dependence 3}
\item{\code{hasHelp}}{Have someone to help with everyday activities}
\item{\code{unmetNeed}}{Need help but has no helper}
}
}

\examples{

# Create adl indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_adl(testSVY)

}
