% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_anthro}
\alias{create_op_anthro}
\title{create_op_anthro}
\usage{
create_op_anthro(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of older people indicators on anthropometry
}
\description{
Create older people indicators dataframe for anthropometry from survey data
collected using the standard RAM-OP questionnaire.
}
\section{Indicators}{
 Anthropometry and screening

\describe{
\item{\code{MUAC}}{Mid-upper arm circumference (mm)}
\item{\code{oedema}}{Bilateral pitting oedema (may not be nutritional)}
\item{\code{screened}}{Either MUAC or oedema checked previously}
}
}

\examples{

# Create anthropometry indicators dataset from RAM-OP survey data collected
# from Addis Ababa, Ethiopia
create_op_anthro(testSVY)

}
