% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_dementia}
\alias{create_op_dementia}
\title{Create older people indicators dataframe for dementia from survey data
collected using the standard RAM-OP questionnaire.}
\usage{
create_op_dementia(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A tibble of older people indicators on dementia
}
\description{
Create older people indicators dataframe for dementia from survey data
collected using the standard RAM-OP questionnaire.
}
\section{Brief Community Screening Instrument for Dementia (CSID)}{


The CSID dementia screening tool is described in:

\cite{Prince M, et al. (2010). A brief dementia screener suitable for use
by non-specialists in resource poor settings - The cross-cultural
derivation and validation of the brief Community Screening Instrument
for Dementia. International Journal of Geriatric Psychiatry, 26(9),
899–907 \doi{10.1002/gps.2622}}

\describe{
\item{\code{DS}}{Probable dementia by CSID screen}
}
}

\examples{
# Create dementia indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_dementia(testSVY)

}
\author{
Mark Myatt
}
