% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_demo}
\alias{create_op_demo}
\title{Create older people indicators dataframe for demography and situation from
survey data collected using the standard RAM-OP questionnaire}
\usage{
create_op_demo(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of older people indicators on demography and situation
}
\description{
Create older people indicators dataframe for demography and situation from
survey data collected using the standard RAM-OP questionnaire
}
\section{Demography and situation}{


\describe{
  \item{\code{psu}}{Primary sampling unit}
  \item{\code{resp1}}{Respondent is SUBJECT}
  \item{\code{resp2}}{Respondent is FAMILY CARER}
  \item{\code{resp3}}{Respondent is OTHER CARER}
  \item{\code{resp4}}{Respondent is OTHER}
  \item{\code{age}}{Age of respondent (years)}
  \item{\code{ageGrp1}}{Age of respondent is between 50 and 59 years}
  \item{\code{ageGrp2}}{Age of respondent is between 50 and 59 years}
  \item{\code{ageGrp3}}{Age of respondent is between 50 and 59 years}
  \item{\code{ageGrp4}}{Age of respondent is between 50 and 59 years}
  \item{\code{ageGrp5}}{Age of respondent is between 50 and 59 years}
  \item{\code{sex1}}{Male}
  \item{\code{sex2}}{Female}
  \item{\code{marital1}}{Marital status = SINGLE}
  \item{\code{marital2}}{Marital status = MARRIED}
  \item{\code{marital3}}{Marital status = LIVING TOGETHER}
  \item{\code{marital4}}{Marital status = DIVORCED}
  \item{\code{marital5}}{Marital status = SEPARATED}
  \item{\code{marital6}}{Marital status = OTHER}
  \item{\code{alone}}{Respondent lives alone}
}
}

\examples{

# Create demography and situation indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_demo(testSVY)

}
\author{
Mark Myatt
}
