% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_disability_females}
\alias{create_op_disability_females}
\title{create_op_disability_females}
\usage{
create_op_disability_females(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of female older people indicators on disability
}
\description{
Create female older people indicators dataframe for disability from survey
data collected using the standard RAM-OP questionnaire
}
\examples{

# Create disability indicators dataset from RAM-OP survey data collected
# from Addis Ababa, Ethiopia
create_op_disability_females(testSVY)

}
