% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_disability_males}
\alias{create_op_disability_males}
\title{create_op_disability_males}
\usage{
create_op_disability_males(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of male older people indicators on disability
}
\description{
Create male older people indicators dataframe for disability from survey
data collected using the standard RAM-OP questionnaire
}
\examples{

# Create disability indicators dataset from RAM-OP survey data collected
# from Addis Ababa, Ethiopia
create_op_disability_males(testSVY)

}
