% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_food}
\alias{create_op_food}
\title{create_op_food}
\usage{
create_op_food(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionniare}
}
\value{
A dataframe of older people indicators on food intake
}
\description{
Create older people indicators for food intake from survey data collected
using the standard RAM-OP questionnaire
}
\examples{

# Create food intake indicators dataset from RAM-OP survey data collected
# from Addis Ababa, Ethiopia
create_op_food(testSVY)

}
