% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_food_females}
\alias{create_op_food_females}
\title{Create female older people indicators dataframe for food intake
from survey data collected using the standard RAM-OP questionnaire}
\usage{
create_op_food_females(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of female older people indicators on food intake
}
\description{
Create female older people indicators dataframe for food intake
from survey data collected using the standard RAM-OP questionnaire
}
\examples{

# Create food intake indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_food_females(testSVY)

}
