% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_hunger}
\alias{create_op_hunger}
\title{create_op_hunger}
\usage{
create_op_hunger(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of older people indicators on housedhold hunger
}
\description{
Create older people indicators for severe food insecurity from survey data
collected using the standard RAM-OP questionnaire
}
\examples{
# Create household hunger indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_hunger(testSVY)

}
