% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_income}
\alias{create_op_income}
\title{Create older people indicators dataframe for income from survey data collected
using the standard RAM-OP questionnaire.}
\usage{
create_op_income(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A tibble of older people indicators on income
}
\description{
Create older people indicators dataframe for income from survey data collected
using the standard RAM-OP questionnaire.
}
\section{Income and income sources}{


\describe{
\item{\code{M1}}{Has a personal income}
\item{\code{M2A}}{Agriculture / fishing / livestock}
\item{\code{M2B}}{Wages / salary}
\item{\code{M2C}}{Sale of charcoal / bricks / &c.}
\item{\code{M2D}}{Trading (e.g. market or shop)}
\item{\code{M2E}}{Investments}
\item{\code{M2F}}{Spending savings / sale of assets}
\item{\code{M2G}}{Charity}
\item{\code{M2H}}{Cash transfer / Social security}
\item{\code{M2I}}{Other}
}
}

\examples{
# Create income indicators dataset from RAM-OP survey data collected from
# Addis Ababa, Ethiopia
create_op_income(testSVY)

}
\author{
Mark Myatt
}
