% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_income_females}
\alias{create_op_income_females}
\title{Create female older people indicators dataframe for income from survey data
collected using the standard RAM-OP questionnaire}
\usage{
create_op_income_females(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A tibble of female older people indicators on income
}
\description{
Create female older people indicators dataframe for income from survey data
collected using the standard RAM-OP questionnaire
}
\examples{

# Create income indicators dataset from RAM-OP survey data collected from
# Addis Ababa, Ethiopia
create_op_income_females(testSVY)

}
