% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_indicators}
\alias{create_op_indicators}
\title{Construct create_op_* expression}
\usage{
create_op_indicators(indicator, svy, sex)
}
\description{
Construct create_op_* expression
}
\keyword{internal}
