% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_mental}
\alias{create_op_mental}
\title{create_op_mental}
\usage{
create_op_mental(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of older people indicators on mental health
}
\description{
Create older people indicators dataframe for mental health from survey data
collected using the standard RAM-OP questionnaire.
}
\section{Indicators}{


#' \strong{K6 Short form psychological distress score}

The K6 score is described in:

\cite{Kessler RC, Andrews G, Colpe LJ, Hiripi E, Mroczek, DK, Normand SLT,
et al. (2002). Short screening scales to monitor population prevalences
and trends in non-specific psychological distress. Psychological
Medicine, 32(6), 959–976 \url{doi:10.1017/S0033291702006074}}

\describe{
\item{\code{K6}}{K6 score}
\item{\code{K6Case}}{K6 score > 12  (in serious psychological distress)}
}
}

\examples{

# Create mental health indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_mental(testSVY)

}
