% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_misc}
\alias{create_op_misc}
\title{Create older people indicators dataframe for miscellaneous indicators from
survey data collected using the standard RAM-OP questionnaire.}
\usage{
create_op_misc(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A tibble of older people miscellaneous indicators
}
\description{
Create older people indicators dataframe for miscellaneous indicators from
survey data collected using the standard RAM-OP questionnaire.
}
\section{Miscellaneous indicators}{


\describe{
\item{\code{chew}}{Problems chewing food}
\item{\code{food}}{Anyone in HH receives a ration}
\item{\code{NFRI}}{Anyone in HH received non-food relief item(s) in previous month}
}
}

\examples{

# Create miscellaneous indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_misc(testSVY)

}
\author{
Mark Myatt
}
