% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_oedema}
\alias{create_op_oedema}
\title{Create older people indicators dataframe for oedema prevalence from survey
data collected using the standard RAM-OP questionnaire.}
\usage{
create_op_oedema(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of older people indicators on oedema prevalence
}
\description{
Create older people indicators dataframe for oedema prevalence from survey
data collected using the standard RAM-OP questionnaire.
}
\section{Indicators}{
 Oedema prevalence

\describe{
\item{\code{oedema}}{Bilateral pitting oedema (may not be nutritional)}
}
}

\examples{

# Create oedema prevalence indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_oedema(testSVY)

}
