% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_oedema_males}
\alias{create_op_oedema_males}
\title{Create male older people indicators dataframe for oedema prevalence from
survey data collected using the standard RAM-OP questionnaire}
\usage{
create_op_oedema_males(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A tibble of male older people indicators on oedema prevalence
}
\description{
Create male older people indicators dataframe for oedema prevalence from
survey data collected using the standard RAM-OP questionnaire
}
\examples{

# Create oedema prevalence indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_oedema_males(testSVY)

}
