% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_screening}
\alias{create_op_screening}
\title{create_op_screening}
\usage{
create_op_screening(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of older people indicators on screening coverage
}
\description{
Create older people indicators dataframe for screening coverage from survey
data collected using the standard RAM-OP questionnaire.
}
\section{Indicators}{
 Screening Coverage

\describe{
\item{\code{screened}}{Either MUAC or oedema checked previously}
}
}

\examples{

# Create screening coverage indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_screening(testSVY)

}
