% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_visual}
\alias{create_op_visual}
\title{create_op_visual}
\usage{
create_op_visual(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of older people indicators on visual impairment
}
\description{
Create older people indicators dataframe for visual impairment from survey
data collected using the standard RAM-OP questionnaire.
}
\section{Indicators}{
 Visual impairment by "Tumbling E" method

The "Tumbling E" method is described in:

\cite{Taylor HR (1978). Applying new design principles to the construction of an
illiterate E Chart. Am J Optom & Physiol Optics 55:348}

\describe{
\item{\code{poorVA}}{Poor visual acuity (correct in < 3 of 4 tests)}
}
}

\examples{

# Create visual impairment indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_visual(testSVY)

}
