% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_wash}
\alias{create_op_wash}
\title{Create older people indicators dataframe for water, sanitation and hygiene
from survey data collected using the standard RAM-OP questionnaire.}
\usage{
create_op_wash(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A tibble of older people indicators on water, sanitation and
hygiene
}
\description{
Create older people indicators dataframe for water, sanitation and hygiene
from survey data collected using the standard RAM-OP questionnaire.
}
\section{Water, sanitation and hygiene (WASH) indicators}{


These are a (core) subset of indicators from:
\url{https://washdata.org/monitoring/methods/core-questions}

\describe{
\item{\code{W1}}{Improved source of drinking water}
\item{\code{W2}}{Safe drinking water (improved source OR adequate treatment)}
\item{\code{W3}}{Improved sanitation facility}
\item{\code{W4}}{Improved non-shared sanitation facility}
}
}

\examples{

# Create water, sanitation and hygiene indicators dataset from RAM-OP survey
# data collected from Addis Ababa, Ethiopia
create_op_wash(testSVY)

}
\author{
Mark Myatt
}
