% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_wash_females}
\alias{create_op_wash_females}
\title{Create female older people indicators dataframe for water, sanitation and
hygiene from survey data collected using the standard RAM-OP questionnaire}
\usage{
create_op_wash_females(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of female older people indicators on water, sanitation
    and hygiene
}
\description{
Create female older people indicators dataframe for water, sanitation and
hygiene from survey data collected using the standard RAM-OP questionnaire
}
\examples{

# Create water, sanitation and hygiene indicators dataset from RAM-OP survey
# data collected from Addis Ababa, Ethiopia
create_op_wash_females(testSVY)

}
