% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_wash_males}
\alias{create_op_wash_males}
\title{Create male older people indicators dataframe for water, sanitation and
hygiene from survey data collected using the standard RAM-OP questionnaire}
\usage{
create_op_wash_males(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A tibble of male older people indicators on water, sanitation and
hygiene
}
\description{
Create male older people indicators dataframe for water, sanitation and
hygiene from survey data collected using the standard RAM-OP questionnaire
}
\examples{

# Create water, sanitation and hygiene indicators dataset from RAM-OP survey
# data collected from Addis Ababa, Ethiopia
create_op_wash_males(testSVY)

}
