% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{createOP}
\alias{createOP}
\title{createOP}
\usage{
createOP(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
Three dataframes of Older People indicators each with 138 variables:
\describe{
\item{\code{indicators.ALL}}{Indicators dataframe for all respondents}
\item{\code{indicators.MALES}}{Indicators dataframe for male respondents}
\item{\code{indicators.FEMALES}}{Indicators dataframe for female repsondents}
}
}
\description{
Create older people indicators dataframe from survey data collected
using the standard RAM-OP questionnaire.
}
\section{Indicators}{


\strong{Demographic}
\describe{
\item{\code{psu}}{Primary sampling unit}
\item{\code{resp1}}{Respondent is SUBJECT}
\item{\code{resp2}}{Respondent is FAMILY CARER}
\item{\code{resp3}}{Respondent is OTHER CARER}
\item{\code{resp4}}{Respondent is OTHER}
\item{\code{age}}{Age of respondent (years)}
\item{\code{ageGrp1}}{Age of respondent is between 50 and 59 years}
\item{\code{ageGrp2}}{Age of respondent is between 50 and 59 years}
\item{\code{ageGrp3}}{Age of respondent is between 50 and 59 years}
\item{\code{ageGrp4}}{Age of respondent is between 50 and 59 years}
\item{\code{ageGrp5}}{Age of respondent is between 50 and 59 years}
\item{\code{sex1}}{Male}
\item{\code{sex2}}{Female}
\item{\code{marital1}}{Marital status = SINGLE}
\item{\code{marital2}}{Marital status = MARRIED}
\item{\code{marital3}}{Marital status = LIVING TOGETHER}
\item{\code{marital4}}{Marital status = DIVORCED}
\item{\code{marital5}}{Marital status = SEPARATED}
\item{\code{marital6}}{Marital status = OTHER}
\item{\code{alone}}{Respondent lives alone}
}


\strong{Dietary intake indicators}

These dietary intake indicators have been purpose-built for older people but
the basic approach used is described in:

\cite{Kennedy G, Ballard T, Dop M C (2011). Guidelines for Measuring Household
and Individual Dietary Diversity. Rome, FAO
\url{http://www.fao.org/docrep/014/i1983e/i1983e00.htm}}

and extended to include indicators of probable adequate intake of a number of
nutrients / micronutrients.

\describe{
\item{\code{MF}}{Meal frequenct}
\item{\code{DDS}}{Dietary Diversity Score (count of 11 groups)}
\item{\code{FG01}}{Cereals}
\item{\code{FG02}}{Roots and tubers}
\item{\code{FG03}}{Fruits and vegetables}
\item{\code{FG04}}{All meat}
\item{\code{FG05}}{Eggs}
\item{\code{FG06}}{Fish}
\item{\code{FG07}}{Legumes, nuts and seeds}
\item{\code{FG08}}{Milk and milk products}
\item{\code{FG09}}{Fats}
\item{\code{FG10}}{Sugar}
\item{\code{FG11}}{Other}
\item{\code{proteinRich}}{Protein rich foods}
\item{\code{pProtein}}{Protein rich plant sources of protein}
\item{\code{aProtein}}{Protein rich animal sources of protein}
\item{\code{pVitA}}{Plant sources of vitamin A}
\item{\code{aVitA}}{Animal sources of vitamin A}
\item{\code{xVitA}}{Any source of vitamin A}
\item{\code{ironRich}}{Iron rich foods}
\item{\code{caRich}}{Calcium rich foods}
\item{\code{znRich}}{Zinc rich foods}
\item{\code{vitB1}}{Vitamin B1-rich foods}
\item{\code{vitB2}}{Vitamin B2-rich foods}
\item{\code{vitB3}}{Vitamin B3-rich foods}
\item{\code{vitB6}}{Vitamin B6-rich foods}
\item{\code{vitB12}}{Vitamin B12-rich foods}
\item{\code{vitBcomplex}}{Vitamin B1/B2/B3/B6/B12-rich foods}
}

\strong{Household Hunger Scale (HHS)}

The HHS is described in:

\cite{Ballard T, Coates J, Swindale A, Deitchler M (2011). Household Hunger
Scale: Indicator Definition and Measurement Guide. Washington DC,
FANTA-2 Bridge, FHI 360
\url{https://www.fantaproject.org/monitoring-and-evaluation/household-hunger-scale-hhs}}

\describe{
\item{\code{HHS1}}{Little or no hunger in household}
\item{\code{HHS2}}{Moderate hunger in household}
\item{\code{HHS3}}{Severe hunger in household}
}

\strong{Katz "Index of Independence in Activities of Daily Living" (ADL) score}

The Katz ADL score is described in:

\cite{Katz S, Ford AB, Moskowitz RW, Jackson BA, Jaffe MW (1963). Studies
of illness in the aged. The Index of ADL: a standardized measure of
biological and psychosocial function. JAMA, 1963, 185(12):914-9
\url{doi:10.1001/jama.1963.03060120024016}}

\cite{Katz S, Down TD, Cash HR, Grotz, RC (1970). Progress in the development
of the index of ADL. The Gerontologist, 10(1), 20-30
\url{doi:10.1093/geront/10.4_Part_1.274}}

\cite{Katz S (1983). Assessing self-maintenance: Activities of daily living,
mobility and instrumental activities of daily living. JAGS, 31(12),
721-726 \url{doi:10.1111/j.1532-5415.1983.tb03391.x}}

\describe{
\item{\code{ADL01}}{Bathing}
\item{\code{ADL02}}{Dressing}
\item{\code{ADL03}}{Toileting}
\item{\code{ADL04}}{Transferring (mobility)}
\item{\code{ADL05}}{Continence}
\item{\code{ADL06}}{Feeding}
\item{\code{scoreADL}}{ADL Score}
\item{\code{classADL1}}{Severity of dependence 1}
\item{\code{classADL2}}{Severity of dependence 2}
\item{\code{classADL3}}{Severity of dependence 3}
\item{\code{hasHelp}}{Have someone to help with everyday activities}
\item{\code{unmetNeed}}{Need help but has no helper}
}

\strong{K6 Short form psychological distress score}

The K6 score is described in:

\cite{Kessler RC, Andrews G, Colpe LJ, Hiripi E, Mroczek, DK, Normand SLT,
et al. (2002). Short screening scales to monitor population prevalences
and trends in non-specific psychological distress. Psychological
Medicine, 32(6), 959–976 \url{doi:10.1017/S0033291702006074}}

\describe{
\item{\code{K6}}{K6 score}
\item{\code{K6Case}}{K6 score > 12  (in serious psychological distress)}
}

\strong{Brief Community Screening Instrument for Dementia (CSID)}

The CSID dementia screening tool is described in:

\cite{Prince M, et al. (2010). A brief dementia screener suitable for use
by non-specialists in resource poor settings - The cross-cultural
derivation and validation of the brief Community Screening Instrument
for Dementia. International Journal of Geriatric Psychiatry, 26(9),
899–907 \url{doi:10.1002/gps.2622}}

\describe{
\item{\code{DS}}{Probable dementia by CSID screen}
}

\strong{Health and health-seeking indicators}

\describe{
\item{\code{H1}}{Chronic condition}
\item{\code{H2}}{Takes drugs regularly for chronic condition}
\item{\code{H31}}{No drugs available}
\item{\code{H32}}{Too expensive / no money}
\item{\code{H33}}{Too old to look for care}
\item{\code{H34}}{Use traditional medicine}
\item{\code{H35}}{Drugs don't help}
\item{\code{H36}}{No-one to help me}
\item{\code{H37}}{No need}
\item{\code{H38}}{Other}
\item{\code{H39}}{No reason given}
\item{\code{H4}}{Recent disease episode}
\item{\code{H5}}{Accessed care for recent disease episode}
\item{\code{H61}}{No drugs available}
\item{\code{H62}}{Too expensive / no money}
\item{\code{H63}}{Too old to look for care}
\item{\code{H64}}{Use traditional medicine}
\item{\code{H65}}{Drugs don't help}
\item{\code{H66}}{No-one to help me}
\item{\code{H67}}{No need}
\item{\code{H68}}{Other}
\item{\code{H69}}{No reason given}
}

\strong{Income and income sources}

\describe{
\item{\code{M1}}{Has a personal income}
\item{\code{M2A}}{Agriculture / fishing / livestock}
\item{\code{M2B}}{Wages / salary}
\item{\code{M2C}}{Sale of charcoal / bricks / &c.}
\item{\code{M2D}}{Trading (e.g. market or shop)}
\item{\code{M2E}}{Investments}
\item{\code{M2F}}{Spending savings / sale of assets}
\item{\code{M2G}}{Charity}
\item{\code{M2H}}{Cash transfer / Social security}
\item{\code{M2I}}{Other}
}

\strong{Water, sanitation and hygiene (WASH) indicators}

These are a (core) subset of indicators from:

\cite{WHO / UNICEF (2006). Core Questions on Drinking-water and Sanitation
for Household Surveys. Geneva, WHO / UNICEF
\url{http://www.who.int/water_sanitation_health/monitoring/household_surveys/en/}}

\describe{
\item{\code{W1}}{Improved source of drinking water}
\item{\code{W2}}{Safe drinking water (improved source OR adequate treatment)}
\item{\code{W3}}{Improved sanitation facility}
\item{\code{W4}}{Improved non-shared sanitation facility}
}

\strong{Anthropometry and screening}

\describe{
\item{\code{MUAC}}{Mid-upper arm circumference (mm)}
\item{\code{oedema}}{Bilateral pitting oedema (may not be nutritional)}
\item{\code{screened}}{Either MUAC or oedema checked previously}
}

\strong{Visual impairment by "Tumbling E" method}

The "Tumbling E" method is described in:

\cite{Taylor HR (1978). Applying new design principles to the construction of an
illiterate E Chart. Am J Optom & Physiol Optics 55:348}

\describe{
\item{\code{poorVA}}{Poor visual acuity (correct in < 3 of 4 tests)}
}

\strong{Miscellaneous indicators}

\describe{
\item{\code{chew}}{Problems chewing food}
\item{\code{food}}{Anyone in HH receives a ration}
\item{\code{NFRI}}{Anyone in HH received non-food relief item(s) in previous month}
}

\strong{Washington Group on Disability}

See:

  \url{http://www.washingtongroup-disability.com}
  \url{https://www.cdc.gov/nchs/washington_group/wg_documents.htm}

for details.

\describe{
\item{\code{wgVisionD0}}{Vision domain 0}
\item{\code{wgVisionD1}}{Vision domain 1}
\item{\code{wgVisionD2}}{Vision domain 2}
\item{\code{wgVisionD3}}{Vision domain 3}
\item{\code{wgHearingD0}}{Hearing domain 0}
\item{\code{wgHearingD1}}{Hearing domain 1}
\item{\code{wgHearingD2}}{Hearing domain 2}
\item{\code{wgHearingD3}}{Hearing domain 3}
\item{\code{wgMobilityD0}}{Mobility domain 0}
\item{\code{wgMobilityD1}}{Mobility domain 1}
\item{\code{wgMobilityD2}}{Mobility domain 2}
\item{\code{wgMobilityD3}}{Mobility domain 3}
\item{\code{wgRememberingD0}}{Remembering domain 0}
\item{\code{wgRememberingD1}}{Remembering domain 1}
\item{\code{wgRememberingD2}}{Remembering domain 2}
\item{\code{wgRememberingD3}}{Remembering domain 3}
\item{\code{wgSelfCareD0}}{Self-care domain 0}
\item{\code{wgSelfCareD1}}{Self-care domain 1}
\item{\code{wgSelfCareD2}}{Self-care domain 2}
\item{\code{wgSelfCareD3}}{Self-care domain 3}
\item{\code{wgCommunicatingD0}}{Communication domain 0}
\item{\code{wgCommunicatingD1}}{Communication domain 1}
\item{\code{wgCommunicatingD2}}{Communication domain 2}
\item{\code{wgCommunicatingD3}}{Communication domain 3}
\item{\code{wgP0}}{Overall 0}
\item{\code{wgP1}}{Overall 1}
\item{\code{wgP2}}{Overall 2}
\item{\code{wgP3}}{Overall 3}
\item{\code{wgPM}}{Any disability}
}
}

\examples{

# Create indicators dataset from RAM-OP survey data collected from
# Addis Ababa, Ethiopia
createOP(testSVY)

}
