% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-estimate_op.R
\name{estimate_op_all}
\alias{estimate_op_all}
\title{Estimate all standard RAM-OP indicators}
\usage{
estimate_op_all(
  x,
  w,
  indicators = c("demo", "anthro", "food", "hunger", "adl", "disability", "mental",
    "dementia", "health", "oedema", "screening", "income", "wash", "visual", "misc"),
  replicates = 399
)
}
\arguments{
\item{x}{Indicators dataset produced by \link{create_op_all} with primary
sampling unit (PSU) in column named \code{PSU}}

\item{w}{A data frame with primary sampling unit (PSU) in column named
\code{psu} and survey weight (i.e. PSU population) in column named
\code{pop}}

\item{indicators}{A character vector of indicator set names to estimate.
Indicator set names are \code{demo}, \code{anthro}, \code{food},
\code{hunger}, \code{disability}, \code{adl}, \code{mental},
\code{dementia}, \code{health}, \code{income}, \code{wash},
\code{visual}, and \code{misc}. Default is all indicator sets.}

\item{replicates}{Number of bootstrap replicates. Default is 399.}
}
\value{
Tibble of boot estimates for all specified standard RAM-OP indicators
}
\description{
Estimate all standard RAM-OP indicators
}
\examples{
estimate_op_all(x = create_op_all(testSVY),
                w = testPSU,
                replicates = 9)

}
