% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-probitBoot.R
\name{estimateProbit}
\alias{estimateProbit}
\title{estimateProbit}
\usage{
estimateProbit(x, w, statistic = bootPROBIT, params = "MUAC",
  outputColumns = "MUAC", replicates = 399)
}
\arguments{
\item{x}{Indicators dataset produced by createOP with primary sampling unit (PSU)
in column named \code{PSU}}

\item{w}{A data frame with primary sampling unit (PSU) in column named \code{psu}
and survey weight (i.e. PSU population) in column named \code{pop}}

\item{statistic}{A function operating on data in \code{x}; fixed to
\code{bootPROBIT} to perform probit estimation}

\item{params}{Parameters (named columns in \code{x}) passed to the function
specified in \code{statistic}; fixed to \code{MUAC} as indicator amenable
to probit estimation}

\item{outputColumns}{Names of columns in output data frame; fixed to \code{MUAC}}

\item{replicates}{Number of bootstrap replicates}
}
\value{
Dataframe of boot estimates using bootPROBIT function
}
\description{
estimateProbit
}
\examples{
  #
  test <- estimateClassic(x = indicators.ALL,
                          w = testPSU,
                          replicates = 9)

  test
}
