% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fullTable}
\alias{fullTable}
\title{Fill out a one-dimensional table to include a specified range of values}
\usage{
fullTable(x, values)
}
\arguments{
\item{x}{A vector to tabulate}

\item{values}{A vector of values to be included in a table}
}
\value{
A one-dimensional table with specified values
}
\description{
Fill out a one-dimensional table to include a specified range of values
}
\author{
Mark Myatt
}
\keyword{internal}
