% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-get_variables.R
\name{get_variables}
\alias{get_variables}
\title{get_variables}
\usage{
get_variables(
  indicators = c("demo", "food", "hunger", "adl", "disability", "mental", "dementia",
    "health", "income", "wash", "anthro", "oedema", "screening", "visual", "misc")
)
}
\arguments{
\item{indicators}{A character vector of indicator set names. Indicator set
names are \code{demo}, \code{food}, \code{hunger}, \code{disability},
\code{adl}, \code{mental}, \code{dementia}, \code{health}, \code{income},
\code{wash}, \code{anthro}, \code{screening}, \code{visual}, and \code{misc}.
Default is all indicator sets.}
}
\value{
A vector of variable names
}
\description{
Function to get appropriate RAM-OP indicator variable names given a
specified indicator set
}
\examples{

get_variables(indicators = c("demo", "food"))

}
