% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.compatible.packages.R
\name{install.compatible.packages}
\alias{install.compatible.packages}
\title{Install the Newest Usable Version of R packages}
\usage{
install.compatible.packages(package_name)
}
\arguments{
\item{package_name}{A character vector of names of packages to install}
}
\description{
Looks through the CRAN archives for the newest published version of
R packages that are compatible with the version of R installed.
}
