% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_newest_usable_version.R
\name{install_newest_usable_version}
\alias{install_newest_usable_version}
\title{Install the Newest Usable Version of an R package}
\usage{
install_newest_usable_version(package)
}
\arguments{
\item{package}{A character vector of length one giving the name of the
package to install}
}
\value{
A list of length two giving the version of the package installed,
  and the newest version of the package published on CRAN
}
\description{
Looks through the CRAN archives for the newest published version of an
R package that is compatible with the version of R installed.
}
