% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{mergeEstimates}
\alias{mergeEstimates}
\title{mergeEstimates}
\usage{
mergeEstimates(x = classicEstimates, y = probitEstimates)
}
\arguments{
\item{x}{Classic estimates data frame}

\item{y}{Probit estimates data frame}
}
\value{
Data frame of combined classic and probit estimates
}
\description{
Concatenate classic and PROBIT estimates into a single data.frame
}
\examples{
  #
  \dontrun{
  test <- mergeEstimates(x = classicEstimates,
                         y = probitEstimates)
  }
}
