% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldr-package.R
\docType{package}
\name{oldr}
\alias{oldr}
\alias{oldr-package}
\title{oldr}
\description{
oldr provides a function to install the latest available version of an R
package for old versions of R.
}
\author{
JB Duck-Mayr
}
