% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-pyramidPlot.R
\name{pyramid.plot}
\alias{pyramid.plot}
\title{Pyramid plot}
\usage{
pyramid.plot(x, g, main = paste("Pyramid plot of",
  deparse(substitute(x)), "by", deparse(substitute(g))),
  xlab = paste(deparse(substitute(g)), "(", levels(g)[1], "/",
  levels(g)[2], ")"), ylab = deparse(substitute(x)))
}
\arguments{
\item{x}{A vector (numeric, factor, character) holding age-groups}

\item{g}{A binary categorical variable (usually sex)}

\item{main}{Plot title}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
Pyramid plot
}
\description{
Pyramid plot
}
\examples{
  #

}
