% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-pyramidPlot.R
\name{pyramid_plot}
\alias{pyramid_plot}
\title{Function to create a pyramid plot}
\usage{
pyramid_plot(
  x,
  g,
  main = paste("Pyramid plot of", deparse(substitute(x)), "by", deparse(substitute(g))),
  xlab = paste(deparse(substitute(g)), "(", levels(g)[1], "/", levels(g)[2], ")"),
  ylab = deparse(substitute(x))
)
}
\arguments{
\item{x}{A vector (numeric, factor, character) holding age-groups}

\item{g}{A binary categorical variable (usually sex)}

\item{main}{Plot title}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
Pyramid plot
}
\description{
Function to create a pyramid plot
}
\examples{
pyramid_plot(
  x = cut(
    testSVY$d2, 
    breaks = seq(from = 60, to = 105, by = 5),
    include.lowest = TRUE
  ),
  g = testSVY$d3
)

}
\author{
Mark Myatt
}
