% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldr.R
\docType{package}
\name{ramOP}
\alias{ramOP}
\alias{ramOP-package}
\title{\code{ramOP} Package}
\description{
An Implementation of Rapid Assessment Method for Older People (RAM-OP) in R
}
\details{
See the README at
\href{https://github.com/validmeasures/ramOP#readme}{GitHub}
}
