% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_adl}
\alias{report_op_adl}
\title{Create a report chunk for activities of daily living indicators}
\usage{
report_op_adl(format = "html")
}
\arguments{
\item{format}{Either html, docx, or odt. Defaults to html.}
}
\value{
A reporting chunk for ADL indicators
}
\description{
Create a report chunk for activities of daily living indicators
}
\examples{
  report_op_adl()

}
