% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_anthro}
\alias{report_op_anthro}
\title{Create a report chunk for anthropometric indicators}
\usage{
report_op_anthro()
}
\value{
A reporting chunk for anhtropometric indicators
}
\description{
Create a report chunk for anthropometric indicators
}
\examples{
  report_op_anthro()

}
