% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-report.R
\name{report_op_dementia}
\alias{report_op_dementia}
\title{Create a report chunk for dementia indicators}
\usage{
report_op_dementia(format = "html")
}
\arguments{
\item{format}{Either html, docx, or odt. Defaults to html.}
}
\value{
A reporting chunk for dementia indicators
}
\description{
Create a report chunk for dementia indicators
}
\examples{
  report_op_dementia()

}
\author{
Ernest Guevarra
}
