% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-report.R
\name{report_op_demo}
\alias{report_op_demo}
\title{Create a report chunk for demography indicators}
\usage{
report_op_demo(format = "html")
}
\arguments{
\item{format}{Either html, docx or odt. Defaults to html.}
}
\value{
A reporting chunk for demographic indicators
}
\description{
Create a report chunk for demography indicators
}
\examples{
  report_op_demo()

}
\author{
Ernest Guevarra
}
