% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_disability}
\alias{report_op_disability}
\title{Create a report chunk for disability indicators}
\usage{
report_op_disability(format = "html")
}
\arguments{
\item{format}{Either html or latex. Defaults to html.}
}
\value{
A reporting chunk for disability indicators
}
\description{
Create a report chunk for disability indicators
}
\examples{
  report_op_disability()

}
