% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_document}
\alias{report_op_document}
\title{Create a report document containing RAM-OP survey results}
\usage{
report_op_document(
  svy,
  estimates,
  filename = "ramOPreport",
  title = "RAM-OP Report",
  output = "html_document"
)
}
\arguments{
\item{svy}{A data.frame collected using the standard RAM-OP questionnaire}

\item{estimates}{A data.frame of RAM-OP results produced by
\link{merge_estimates} function.}

\item{filename}{Filename for output document. Can be specified as a path to a
specific directory where to output report document}

\item{title}{Title of report}

\item{output}{Type of output. One of two choices: 1) \code{html_document}
to produce and HTML output; or 2) \code{pdf_document} to produce a PDF
output. Default is \code{html_document}}
}
\value{
A document of specified output type in the working directory or
  if filename is a path, to a specified directory.
}
\description{
Create a report document containing RAM-OP survey results
}
\examples{
  #
  classicResults <- estimate_classic(x = create_op_all(testSVY),
                                     w = testPSU,
                                     replicates = 9)

  probitResults <- estimate_probit(x = create_op_all(testSVY),
                                   w = testPSU,
                                   replicates = 9)

  resultsDF <- merge_estimates(x = classicResults, y = probitResults)

  report_op_document(svy = testSVY,
                     estimates = resultsDF,
                     filename = paste(tempdir(), report, sep = "/"))

}
