% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_food}
\alias{report_op_food}
\title{Create a report chunk for food indicators}
\usage{
report_op_food()
}
\value{
A reporting chunk for food indicators
}
\description{
Create a report chunk for food indicators
}
\examples{
  report_op_food()

}
