% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-report.R
\name{report_op_health}
\alias{report_op_health}
\title{Create a report chunk for health and health-seeking behaviour indicators}
\usage{
report_op_health(format = "html")
}
\arguments{
\item{format}{Either html, docx, or odt. Defaults to html.}
}
\value{
A reporting chunk for health and health-seeking behaviour indicators
}
\description{
Create a report chunk for health and health-seeking behaviour indicators
}
\examples{
  report_op_health()

}
\author{
Ernest Guevarra
}
