% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_html}
\alias{report_op_html}
\title{Create an HTML report document containing RAM-OP survey results}
\usage{
report_op_html(
  estimates,
  svy,
  filename = "ramOPreport",
  title = "RAM-OP Report"
)
}
\arguments{
\item{estimates}{A data.frame of RAM-OP results produced by
\link{merge_estimates} function.}

\item{svy}{A data.frame collected using the standard RAM-OP questionnaire}

\item{filename}{Filename for output document. Can be specified as a path to a
specific directory where to output report document}

\item{title}{Title of report}
}
\value{
An HTML document in the working directory or if filename is a path,
  to a specified directory.
}
\description{
Create an HTML report document containing RAM-OP survey results
}
\examples{
  #
  classicResults <- estimate_classic(x = create_op_all(testSVY),
                                     w = testPSU,
                                     replicates = 9)

  probitResults <- estimate_probit(x = create_op_all(testSVY),
                                   w = testPSU,
                                   replicates = 9)

  resultsDF <- merge_estimates(x = classicResults, y = probitResults)

  report_op_html(svy = testSVY,
                 estimates = resultsDF,
                 filename = paste(tempdir(), "report", sep = "/"))

}
