% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_hunger}
\alias{report_op_hunger}
\title{Create a report chunk for activities of food security indicators}
\usage{
report_op_hunger(format = "html")
}
\arguments{
\item{format}{Either html, docx, or odt. Defaults to html.}
}
\value{
A reporting chunk for food security indicators
}
\description{
Create a report chunk for activities of food security indicators
}
\examples{
  report_op_hunger()

}
\author{
Ernest Guevarra
}
