% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-report.R
\name{report_op_mental}
\alias{report_op_mental}
\title{Create a report chunk for mental health indicators}
\usage{
report_op_mental(format = "html")
}
\arguments{
\item{format}{Either html, docx, or odt. Defaults to html.}
}
\value{
A reporting chunk for mental health indicators
}
\description{
Create a report chunk for mental health indicators
}
\examples{
  report_op_mental()

}
\author{
Ernest Guevarra
}
