% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_misc}
\alias{report_op_misc}
\title{Create a report chunk for miscellaneous indicators}
\usage{
report_op_misc()
}
\value{
A reporting chunk for miscellaneous indicators
}
\description{
Create a report chunk for miscellaneous indicators
}
\examples{
  report_op_misc()

}
