% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_oedema}
\alias{report_op_oedema}
\title{Create a report chunk for oedema}
\usage{
report_op_oedema()
}
\value{
A reporting chunk for oedema indicators
}
\description{
Create a report chunk for oedema
}
\examples{
  report_op_oedema()

}
