% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-report.R
\name{report_op_table}
\alias{report_op_table}
\title{Create table of RAM-OP results}
\usage{
report_op_table(estimates, filename = paste(tempdir(), "ramOP", sep = "/"))
}
\arguments{
\item{estimates}{A data.frame of RAM-OP results produced by
\link{merge_estimates} function.}

\item{filename}{Prefix to append to report output filename. Can be specified
as a path to a specific directory where to output tabular results CSV file.
Defaults to a path to a temporary directory with a filename starting with
\code{ramOP}.}
}
\value{
Report of tabulated estimated results saved in CSV format in current
working directory or in the specified path
}
\description{
Create table of RAM-OP results
}
\examples{
  ##
  x <- estimate_classic(x = create_op(testSVY),
                        w = testPSU,
                        replicates = 9)
  y <- estimate_probit(x = create_op(testSVY),
                       w = testPSU,
                       replicates = 9)
  z <- merge_estimates(x, y, prop2percent = TRUE)
  report_op_table(z)

}
\author{
Mark Myatt
}
