% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{reportOP}
\alias{reportOP}
\title{reportOP}
\usage{
reportOP(FILENAME)
}
\arguments{
\item{FILENAME}{Prefix to append to report output filename}
}
\value{
Report of tabulated estimated results saved in CSV format in current
    working directory
}
\description{
Create table of RAM-OP results
}
\examples{
  #
  \dontrun{
    reportOP("TEST")
  }
}
