% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldr.R
\docType{data}
\name{testPSU}
\alias{testPSU}
\title{RAM-OP Population Dataset}
\format{A data frame with 2 columns and 16 rows:
\describe{
\item{\code{psu}}{The PSU identifier. This must use the same coding system
  used to identify the PSUs that is used in the main RAM-OP dataset}
\item{\code{pop}}{The population of the PSU}
}
The PSU dataset is used during data analysis to weight data by PSU
population.}
\usage{
testPSU
}
\description{
This is a short and narrow file with one record per PSU and just two
variables
}
\keyword{datasets}
