% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-chart.R
\name{chart_age}
\alias{chart_age}
\title{Age by sex (pyramid plot)}
\usage{
chart_age(
  x,
  save_chart = TRUE,
  filename = paste(tempdir(), "populationPyramid", sep = "/")
)
}
\arguments{
\item{x}{Indicators dataset produced by \code{\link[=create_op]{create_op()}}}

\item{save_chart}{Logical. Should chart be saved? Default is TRUE.}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory. Default is a path to
a temporary directory and a filename starting with
\code{populationPyramid}. Ignored if \code{save_chart} is FALSE.}
}
\value{
Age by sex pyramid plot in PNG format saved in the current working
directory or in a specified directory if \code{filename} is a path unless
when \code{save_chart} is FALSE in which case the plot is shown on current
graphics device
}
\description{
A wrapper function to the \code{\link[=pyramid.plot]{pyramid.plot()}} function to create an age by
sex pyramid plot
}
\examples{
  # Create age by sex pyramid plot using indicators.ALL dataset
  chart_age(x = indicators.ALL)

}
