% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-chart.R
\name{chart_income}
\alias{chart_income}
\title{Chart income indicators}
\usage{
chart_income(x.male, x.female, save_chart = TRUE, filename = NULL)
}
\arguments{
\item{x.male}{Male subset of indicator dataset}

\item{x.female}{Female subset of indicator dataset}

\item{save_chart}{Logical. Should chart be saved? Default is TRUE.}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory. Ignored if
\code{save_chart} is FALSE.}
}
\value{
Bar chart of sources of income by sex in PNG format saved in current
  working directory or in a specified directory if \code{filename} is a path
  unless when \code{save_chart} is FALSE in which case chart is shown in
  current graphics device.
}
\description{
Chart income indicators
}
\examples{
  # Create chart using indicators.FEMALES and indicators.MALES
  # dataset
  chart_income(x.male = indicators.MALES,
               x.female = indicators.FEMALES,
               filename = paste(tempdir(), "TEST", sep = "/"))

}
