% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-chart.R
\name{chart_k6}
\alias{chart_k6}
\title{Distribution of K6 (overall and by sex)}
\usage{
chart_k6(x, save_chart = TRUE, filename = paste(tempdir(), "chart", sep = "/"))
}
\arguments{
\item{x}{Indicators dataset produced by \code{\link[=create_op]{create_op()}}}

\item{save_chart}{Logical. Should chart be saved? Default is TRUE.}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory. Defaults to a path to
a temporary directory and a filename starting with \code{chart}. Ignored if
\code{save_chart} is FALSE.}
}
\value{
Histogram of K6 score in PNG format saved in current
working directory or in a specified directory if \code{filename} is a
path unless when \code{save_chart} is FALSE in which case chart is shown
in current graphics device.
}
\description{
Distribution of K6 (overall and by sex)
}
\examples{
  # Create chart using indicators.ALL dataset
  chart_k6(x = indicators.ALL)

}
