% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-chart.R
\name{chart_muac}
\alias{chart_muac}
\title{Distribution of MUAC (overall and by sex)}
\usage{
chart_muac(
  x,
  save_chart = TRUE,
  filename = paste(tempdir(), "chart", sep = "/")
)
}
\arguments{
\item{x}{Indicators dataset produced by \link{create_op_all}}

\item{save_chart}{Logical. Should chart be saved? Default is TRUE.}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory. Default is a path to
a temporary directory and a filename starting with \code{chart}. Ignored
if \code{save_chart} is FALSE.}
}
\value{
Histogram of MUAC distribution in PNG format and saved in the current
working directory or in a specified directory if \code{filename} is a
path unless when \code{save_chart} is FALSE in which case chart is
shown on current graphics device.
}
\description{
Distribution of MUAC (overall and by sex)
}
\examples{
  # Create MUAC histogram using indicators.ALL dataset
  chart_muac(x = indicators.ALL)

}
