% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-chart.R
\name{chart_wash}
\alias{chart_wash}
\title{Chart WASH indicators}
\usage{
chart_wash(
  x,
  save_chart = TRUE,
  filename = paste(tempdir(), "chart", sep = "/")
)
}
\arguments{
\item{x}{Indicators dataset produced by \link{create_op_all}}

\item{save_chart}{Logical. Should chart be saved? Default is TRUE.}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory. Defaults to a path to
a temporary directory and a filename starting with \code{chart}. Ignored if
\code{save_chart} is FALSE.}
}
\value{
Bar plot of ADL in PNG format saved in current working directory
or in a specified directory if \code{filename} is a path unless when
\code{save_chart} is FALSE in which case chart is shown in current
graphic device
}
\description{
Chart WASH indicators
}
\examples{
  # Create chart using indicators.ALL dataset
  chart_wash(x = indicators.ALL)

}
