% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-chart.R
\name{chart_wg}
\alias{chart_wg}
\title{Chart disability (Washington Group - WG) indicators}
\usage{
chart_wg(x, save_chart = TRUE, filename = paste(tempdir(), "chart", sep = "/"))
}
\arguments{
\item{x}{Indicators dataset produced by \code{\link[=create_op]{create_op()}}}

\item{save_chart}{Logical. Should chart be saved? Default is TRUE.}

\item{filename}{Prefix to add to output chart filename or a directory
path to save output to instead of working directory. Defaults to a path to
a working directory and a filename starting with \code{chart}. Ignored if
\code{save_chart} is FALSE.}
}
\value{
Bar plot of Disability Score in PNG format saved in current working
directory or in a specified directory if \code{filename} is a path
unless when \code{save_chart} is FALSE in which case chart is shown in
current graphic device.
}
\description{
Chart disability (Washington Group - WG) indicators
}
\examples{
  # Create chart using indicators.ALL dataset
  chart_wg(x = indicators.ALL)

}
