% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chartCSID}
\alias{chartCSID}
\title{Chart dementia screen (CSID) indicators}
\usage{
chartCSID(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Bar plot of CSID in PNG format saved in current working directory
}
\description{
Chart dementia screen (CSID) indicators
}
\examples{
  # Create chart using \code{indicators.ALL} dataset
  chartCSID(x = indicators.ALL, filename = paste(tempdir(), "TEST", sep = "/"))
}
