% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-chart.R
\name{chartHHS}
\alias{chartHHS}
\title{Chart household hunger scale (HHS) indicators}
\usage{
chartHHS(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Bar plot of HHS in PNG format saved in current working directory
}
\description{
Chart household hunger scale (HHS) indicators
}
\examples{
  # Create chart using \code{indicators.ALL} dataset
  chartHHS(x = indicators.ALL, filename = paste(tempdir(), "TEST", sep = "/"))
}
