% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-chart.R
\name{chartWASH}
\alias{chartWASH}
\title{chartWASH}
\usage{
chartWASH(x, filename)
}
\arguments{
\item{x}{Indicators dataset produced by createOP}

\item{filename}{Prefix to add to output chart filename}
}
\value{
Bar plot of ADL in PNG format saved in current working directory
}
\description{
WASH indicators
}
\examples{
  # Create chart using \\code{indicators.ALL} dataset
  chartWASH(x = indicators.ALL, filename = "TEST")
}
