% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_adl_males}
\alias{create_op_adl_males}
\title{Create male older people indicators dataframe for activities of daily
living from survey data collected using the standard RAM-OP questionnaire}
\usage{
create_op_adl_males(svy)
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of male older people indicators on activities of daily
living
}
\description{
Create male older people indicators dataframe for activities of daily
living from survey data collected using the standard RAM-OP questionnaire
}
\examples{

# Create activities of daily living indicators dataset from RAM-OP survey
# data collected from Addis Ababa, Ethiopia
create_op_adl_males(testSVY)

}
