% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_all}
\alias{create_op_all}
\title{create_op_all}
\usage{
create_op_all(svy, indicators = c("demo", "food", "hunger", "disability",
  "adl", "mental", "dementia", "health", "income", "wash", "anthro",
  "visual", "misc"))
}
\arguments{
\item{svy}{A dataframe collected using the standard RAM-OP questionnaire}

\item{indicators}{A character vector of indicator names}
}
\value{
A dataframe of older people indicators
}
\description{
Create older people indicators dataframe from survey data collected
using the standard RAM-OP questionnaire.
}
\examples{

create_op_all(svy = testSVY)

}
