% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-opIndicators.R
\name{create_op_food}
\alias{create_op_food}
\title{Create older people indicators for food intake from survey data collected
using the standard RAM-OP questionnaire}
\usage{
create_op_food(svy)
}
\arguments{
\item{svy}{A data.frame collected using the standard RAM-OP questionnaire}
}
\value{
A dataframe of older people indicators on food intake
}
\description{
Create older people indicators for food intake from survey data collected
using the standard RAM-OP questionnaire
}
\section{Dietary intake indicators}{


These dietary intake indicators have been purpose-built for older people but
the basic approach used is described in:

\cite{Kennedy G, Ballard T, Dop M C (2011). Guidelines for Measuring Household
and Individual Dietary Diversity. Rome, FAO
\url{http://www.fao.org/docrep/014/i1983e/i1983e00.htm}}

and extended to include indicators of probable adequate intake of a number of
nutrients / micronutrients.

\describe{
\item{\code{MF}}{Meal frequency}
\item{\code{DDS}}{Dietary Diversity Score (count of 11 groups)}
\item{\code{FG01}}{Cereals}
\item{\code{FG02}}{Roots and tubers}
\item{\code{FG03}}{Fruits and vegetables}
\item{\code{FG04}}{All meat}
\item{\code{FG05}}{Eggs}
\item{\code{FG06}}{Fish}
\item{\code{FG07}}{Legumes, nuts and seeds}
\item{\code{FG08}}{Milk and milk products}
\item{\code{FG09}}{Fats}
\item{\code{FG10}}{Sugar}
\item{\code{FG11}}{Other}
\item{\code{proteinRich}}{Protein rich foods}
\item{\code{pProtein}}{Protein rich plant sources of protein}
\item{\code{aProtein}}{Protein rich animal sources of protein}
\item{\code{pVitA}}{Plant sources of vitamin A}
\item{\code{aVitA}}{Animal sources of vitamin A}
\item{\code{xVitA}}{Any source of vitamin A}
\item{\code{ironRich}}{Iron rich foods}
\item{\code{caRich}}{Calcium rich foods}
\item{\code{znRich}}{Zinc rich foods}
\item{\code{vitB1}}{Vitamin B1-rich foods}
\item{\code{vitB2}}{Vitamin B2-rich foods}
\item{\code{vitB3}}{Vitamin B3-rich foods}
\item{\code{vitB6}}{Vitamin B6-rich foods}
\item{\code{vitB12}}{Vitamin B12-rich foods}
\item{\code{vitBcomplex}}{Vitamin B1/B2/B3/B6/B12-rich foods}
}
}

\examples{
# Create food intake indicators dataset from RAM-OP survey data collected
# from Addis Ababa, Ethiopia
create_op_food(testSVY)

}
\author{
Mark Myatt
}
