% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-classicBoot.R
\name{estimateClassic}
\alias{estimateClassic}
\title{estimateClassic}
\usage{
estimateClassic(x, w, statistic = bootClassic, params = c("resp1",
  "resp2", "resp3", "resp4", "age", "ageGrp1", "ageGrp2", "ageGrp3",
  "ageGrp4", "ageGrp5", "sex1", "sex2", "marital1", "marital2", "marital3",
  "marital4", "marital5", "marital6", "alone", "MF", "DDS", "FG01", "FG02",
  "FG03", "FG04", "FG05", "FG06", "FG07", "FG08", "FG09", "FG10", "FG11",
  "proteinRich", "pProtein", "aProtein", "pVitA", "aVitA", "xVitA",
  "ironRich", "caRich", "znRich", "vitB1", "vitB2", "vitB3", "vitB6",
  "vitB12", "vitBcomplex", "HHS1", "HHS2", "HHS3", "ADL01", "ADL02",
  "ADL03",      "ADL04", "ADL05", "ADL06", "scoreADL", "classADL1",
  "classADL2", "classADL3", "hasHelp", "unmetNeed", "K6", "K6Case", "DS",
  "H1", "H2", "H31", "H32", "H33", "H34", "H35", "H36", "H37", "H38",
  "H39", "H4", "H5", "H61", "H62", "H63", "H64", "H65", "H66", "H67",
  "H68", "H69", "M1", "M2A", "M2B", "M2C", "M2D", "M2E", "M2F", "M2G",
  "M2H", "M2I", "W1", "W2", "W3", "W4", "MUAC", "oedema", "screened",
  "poorVA", "chew", "food", "NFRI", "wgVisionD0", "wgVisionD1",
  "wgVisionD2", "wgVisionD3", "wgHearingD0",      "wgHearingD1",
  "wgHearingD2", "wgHearingD3", "wgMobilityD0", "wgMobilityD1",
  "wgMobilityD2", "wgMobilityD3", "wgRememberingD0", "wgRememberingD1",
  "wgRememberingD2", "wgRememberingD3", "wgSelfCareD0", "wgSelfCareD1",
  "wgSelfCareD2", "wgSelfCareD3", "wgCommunicatingD0", "wgCommunicatingD1",
  "wgCommunicatingD2", "wgCommunicatingD3", "wgP0", "wgP1", "wgP2", "wgP3",
  "wgPM"), outputColumns = c("resp1", "resp2", "resp3", "resp4", "age",
  "ageGrp1", "ageGrp2", "ageGrp3", "ageGrp4", "ageGrp5", "sex1", "sex2",
  "marital1", "marital2", "marital3", "marital4", "marital5", "marital6",
  "alone", "MF", "DDS", "FG01", "FG02", "FG03", "FG04", "FG05", "FG06",
  "FG07", "FG08", "FG09", "FG10", "FG11", "proteinRich", "pProtein",
  "aProtein", "pVitA", "aVitA", "xVitA", "ironRich", "caRich", "znRich",
  "vitB1", "vitB2", "vitB3", "vitB6", "vitB12", "vitBcomplex", "HHS1",
  "HHS2", "HHS3", "ADL01", "ADL02", "ADL03",      "ADL04", "ADL05",
  "ADL06", "scoreADL", "classADL1", "classADL2", "classADL3", "hasHelp",
  "unmetNeed", "K6", "K6Case", "DS", "H1", "H2", "H31", "H32", "H33",
  "H34", "H35", "H36", "H37", "H38", "H39", "H4", "H5", "H61", "H62",
  "H63", "H64", "H65", "H66", "H67", "H68", "H69", "M1", "M2A", "M2B",
  "M2C", "M2D", "M2E", "M2F", "M2G", "M2H", "M2I", "W1", "W2", "W3", "W4",
  "MUAC", "oedema", "screened", "poorVA", "chew", "food", "NFRI",
  "wgVisionD0", "wgVisionD1", "wgVisionD2", "wgVisionD3", "wgHearingD0",   
    "wgHearingD1", "wgHearingD2", "wgHearingD3", "wgMobilityD0",
  "wgMobilityD1", "wgMobilityD2", "wgMobilityD3", "wgRememberingD0",
  "wgRememberingD1", "wgRememberingD2", "wgRememberingD3", "wgSelfCareD0",
  "wgSelfCareD1", "wgSelfCareD2", "wgSelfCareD3", "wgCommunicatingD0",
  "wgCommunicatingD1", "wgCommunicatingD2", "wgCommunicatingD3", "wgP0",
  "wgP1", "wgP2", "wgP3", "wgPM"), replicates = 399)
}
\arguments{
\item{x}{Indicators dataset produced by createOP with primary sampling unit (PSU)
in column named \code{PSU}}

\item{w}{A data frame with primary sampling unit (PSU) in column named \code{psu}
and survey weight (i.e. PSU population) in column named \code{pop}}

\item{statistic}{A function operating on data in \code{x}; fixed to
\code{bootClassic} function for means}

\item{params}{Parameters (named columns in \code{x}) passed to the function
specified in \code{statistic}; defaults to all variables in indicators
data frame except for MUAC}

\item{outputColumns}{Names of columns in output data frame; ; defaults to all
variables in indicators data frame except for MUAC}

\item{replicates}{Number of bootstrap replicates}
}
\value{
Dataframe of boot estimates using bootClassic mean function
}
\description{
estimateClassic
}
\examples{
  #
  test <- estimateClassic(x = indicators.ALL,
                          w = testPSU,
                          replicates = 9)

  test
}
