% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fullTable.R
\name{fullTable}
\alias{fullTable}
\title{Fill out a one-dimensional table to include a specified range of values}
\usage{
fullTable(x, values)
}
\arguments{
\item{x}{A vector to tabulate}

\item{values}{A vector of values to be included in a table}
}
\value{
A one-dimensional table with specified values
}
\description{
Fill out a one-dimensional table to include a specified range of values
}
\examples{

  xTable <- fullTable(x = sample(x = 5,
                                 size = 100,
                                 replace = TRUE),
                      values = 1:5)
  xTable
}
\author{
Mark Myatt
}
