% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Get appropriate RAM-OP indicator variable names given a specified indicator
set}
\usage{
get_variables(
  indicators = c("demo", "food", "hunger", "adl", "disability", "mental", "dementia",
    "health", "income", "wash", "anthro", "oedema", "screening", "visual", "misc")
)
}
\arguments{
\item{indicators}{A character vector of indicator set names. Indicator set
names are \emph{"demo"}, \emph{"food"}, \emph{"hunger"}, \emph{"disability"}, \emph{"adl"},
\emph{"mental"}, \emph{"dementia"}, \emph{"health"}, \emph{"income"}, \emph{"wash"}, \emph{"anthro"},
\emph{"screening"}, \emph{"visual"}, and \emph{"misc"}. Default is all indicator sets.}
}
\value{
A vector of variable names
}
\description{
Get appropriate RAM-OP indicator variable names given a specified indicator
set
}
\examples{

get_variables(indicators = c("demo", "food"))

}
