% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.compatible.packages.R
\name{install.compatible.packages}
\alias{install.compatible.packages}
\title{Install the Newest Usable Version of R packages}
\usage{
install.compatible.packages(package_name, R_version = "installed_version",
  lib = .libPaths()[1])
}
\arguments{
\item{package_name}{A character vector of names of packages to install}

\item{R_version}{A character vector giving the R version(s) to attempt
installation for. The default, "installed_version", will attempt
installation for the version of R from which the function was called.
Any entry that does not contain a version number (e.g. "3.4.3") will
be ignored.}

\item{lib}{A character vector giving the directories to install the packages
to (see \code{\link[utils]{install.packages}}). The vector should be
either length one or of the same length as R_version; the first element
will be used for installing packages for the first R version supplied,
the second for the second version, etc.; if only one lib is supplied,
all packages for all versions will be installed in that directory,
so if installation is done for multiple R versions, the only surviving
installations will be for the last supplied element of R_version.
The default is the first element of \code{\link[base]{.libPaths}}().}
}
\description{
Looks through the CRAN archives for the newest published version of
R packages that are compatible with the version of R installed.
}
