% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-oldr.R
\docType{package}
\name{oldr}
\alias{oldr}
\alias{oldr-package}
\title{\code{oldr} Package}
\description{
An Implementation of Rapid Assessment Method for Older People (RAM-OP) in R
}
\details{
See the README at
\href{https://github.com/validmeasures/oldr#readme}{GitHub}
}
