% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-report.R
\name{report_op_table}
\alias{report_op_table}
\alias{report_op_demo}
\alias{report_op_food}
\alias{report_op_hunger}
\alias{report_op_disability}
\alias{report_op_adl}
\alias{report_op_mental}
\alias{report_op_dementia}
\alias{report_op_health}
\alias{report_op_oedema}
\alias{report_op_anthro}
\alias{report_op_screen}
\alias{report_op_visual}
\alias{report_op_income}
\alias{report_op_wash}
\alias{report_op_misc}
\title{Create table and report chunk of RAM-OP results}
\usage{
report_op_table(estimates, filename = paste(tempdir(), "ramOP", sep = "/"))

report_op_demo(format = "html")

report_op_food(format = "html")

report_op_hunger(format = "html")

report_op_disability(format = "html")

report_op_adl(format = "html")

report_op_mental(format = "html")

report_op_dementia(format = "html")

report_op_health(format = "html")

report_op_oedema(format = "html")

report_op_anthro(format = "html")

report_op_screen(format = "html")

report_op_visual(format = "html")

report_op_income(format = "html")

report_op_wash(format = "html")

report_op_misc(format = "html")
}
\arguments{
\item{estimates}{A data.frame of RAM-OP results produced by \code{\link[=merge_op]{merge_op()}}.}

\item{filename}{Prefix to append to report output filename. Can be specified
as a path to a specific directory where to output tabular results CSV file.
Defaults to a path to a temporary directory with a filename starting with
\emph{ramOP}.}

\item{format}{Either \emph{"html"}, \emph{"docx"} or \emph{"odt"}. Defaults to \emph{"html"}.}
}
\value{
Report of tabulated estimated results saved in CSV format in current
working directory or in the specified path or a reporting chunk for
specific indicators.
}
\description{
Create table and report chunk of RAM-OP results
}
\examples{
##
x <- estimate_classic(
  x = create_op(testSVY), w = testPSU, replicates = 9
)

y <- estimate_probit(
  x = create_op(testSVY), w = testPSU, replicates = 9
)

z <- merge_op(x, y, prop2percent = TRUE)
report_op_table(z)

report_op_demo()
report_op_hunger()
report_op_food()
report_op_disability()

}
\author{
Mark Myatt and Ernest Guevarra
}
