% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-report.R
\name{report_op_anthro}
\alias{report_op_anthro}
\title{Create a report chunk for anthropometric indicators}
\usage{
report_op_anthro(format = "html")
}
\arguments{
\item{format}{Either html, docx, or odt. Defaults to html.}
}
\value{
A reporting chunk for anthropometric indicators
}
\description{
Create a report chunk for anthropometric indicators
}
\examples{
  report_op_anthro()

}
\author{
Ernest Guevarra
}
