% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-report.R
\name{report_op_food}
\alias{report_op_food}
\title{Create a report chunk for food indicators}
\usage{
report_op_food(format = "html")
}
\arguments{
\item{format}{Either html, docx, or odt. Defaults to html.}
}
\value{
A reporting chunk for food indicators
}
\description{
Create a report chunk for food indicators
}
\examples{
  report_op_food()

}
\author{
Ernest Guevarra
}
