% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_income}
\alias{report_op_income}
\title{Create a report chunk for income}
\usage{
report_op_income(format = "html")
}
\arguments{
\item{format}{Either html or latex. Defaults to html.}
}
\value{
A reporting chunk for income
}
\description{
Create a report chunk for income
}
\examples{
  report_op_income()

}
