% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-report.R
\name{report_op_misc}
\alias{report_op_misc}
\title{Create a report chunk for miscellaneous indicators}
\usage{
report_op_misc(format = "html")
}
\arguments{
\item{format}{Either html, docx or odt. Defaults to html.}
}
\value{
A reporting chunk for miscellaneous indicators
}
\description{
Create a report chunk for miscellaneous indicators
}
\examples{
  report_op_misc()

}
\author{
Ernest Guevarra
}
