% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_screen}
\alias{report_op_screen}
\title{Create a report chunk for screening indicators}
\usage{
report_op_screen(format = "html")
}
\arguments{
\item{format}{Either html, docx, or odt. Defaults to html.}
}
\value{
A reporting chunk for screening indicators
}
\description{
Create a report chunk for screening indicators
}
\examples{
  report_op_screen()

}
\author{
Ernest Guevarra
}
