% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-report.R
\name{report_op_table}
\alias{report_op_table}
\title{Create table of RAM-OP results}
\usage{
report_op_table(estimates, filename)
}
\arguments{
\item{estimates}{A data.frame of RAM-OP results produced by \link{merge_estimates}
function.}

\item{filename}{Prefix to append to report output filename. Can be specified
as a path to a specific directory where to output tabular results CSV file}
}
\value{
Report of tabulated estimated results saved in CSV format in current
    working directory or in the specified path
}
\description{
Create table of RAM-OP results
}
\examples{
  #
  \dontrun{
    report_op_table("TEST")
  }

}
