% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_visual}
\alias{report_op_visual}
\title{Create a report chunk for visual acuity}
\usage{
report_op_visual(format = "html")
}
\arguments{
\item{format}{Either html, docx, or odt. Defaults to html.}
}
\value{
A reporting chunk for visual acuity
}
\description{
Create a report chunk for visual acuity
}
\examples{
  report_op_visual()

}
\author{
Ernest Guevarra
}
