% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-report.R
\name{report_op_wash}
\alias{report_op_wash}
\title{Create a report chunk for water, sanitation and hygiene}
\usage{
report_op_wash(format = "html")
}
\arguments{
\item{format}{Either html or latex. Defaults to html.}
}
\value{
A reporting chunk for water, sanitation and hygiene
}
\description{
Create a report chunk for water, sanitation and hygiene
}
\examples{
  report_op_wash()

}
